tests/WebPlatformSubset/contracts/observed-native-input-event-primitives.html

<!doctype html>
<meta charset="utf-8">
<title>Observed native pointer and mouse event lifecycle primitives</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<style>
  html,
  body {
    margin: 0;
    width: 360px;
    height: 140px;
  }

  #target,
  #outside {
    position: fixed;
    top: 20px;
    width: 100px;
    height: 40px;
  }

  #target { left: 20px; }
  #outside { left: 220px; }

  #propagation-parent {
    position: fixed;
    left: 20px;
    top: 85px;
    width: 160px;
    height: 40px;
  }
</style>
<button id="target">Target</button>
<button id="outside">Outside</button>
<div id="propagation-parent">
  <button id="propagation-target">Propagation target</button>
</div>
<script>
  const target = document.getElementById("target");
  const outside = document.getElementById("outside");
  const observed = [];
  const types = [
    "pointerover", "pointerenter", "pointermove", "pointerdown", "pointerup",
    "pointerout", "pointerleave", "mouseover", "mouseenter", "mousemove",
    "mousedown", "mouseup", "mouseout", "mouseleave", "click"
  ];

  for (const type of types) {
    target.addEventListener(type, event => observed.push({
      type,
      target: event.target,
      relatedTarget: event.relatedTarget,
      clientX: event.clientX,
      clientY: event.clientY,
      offsetX: event.offsetX,
      offsetY: event.offsetY,
      movementX: event.movementX,
      movementY: event.movementY,
      button: event.button,
      buttons: event.buttons,
      which: event.which,
      detail: event.detail,
      pointerType: event.pointerType,
      isPrimary: event.isPrimary
    }));
  }

  const events = type => observed.filter(event => event.type === type);
  const count = type => events(type).length;

  promise_test(async () => {
    await new test_driver.Actions().pointerMove(0, 0, { origin: outside }).send();
    await new test_driver.Actions().pointerMove(0, 0, { origin: target }).send();

    for (const type of ["pointerover", "pointerenter", "mouseover", "mouseenter",
                        "pointermove", "mousemove"]) {
      assert_equals(count(type), 1, `${type} must be dispatched exactly once on entry`);
      assert_equals(events(type)[0].target, target, `${type}.target`);
    }
    assert_less_than(observed.findIndex(event => event.type === "pointerover"),
      observed.findIndex(event => event.type === "pointerenter"),
      "pointerover precedes pointerenter");
    assert_less_than(observed.findIndex(event => event.type === "pointerenter"),
      observed.findIndex(event => event.type === "pointermove"),
      "pointer boundary events precede pointermove");
    assert_less_than(observed.findIndex(event => event.type === "mouseover"),
      observed.findIndex(event => event.type === "mouseenter"),
      "mouseover precedes mouseenter");
    assert_less_than(observed.findIndex(event => event.type === "mouseenter"),
      observed.findIndex(event => event.type === "mousemove"),
      "mouse boundary events precede mousemove");

    const move = events("pointermove")[0];
    assert_equals(move.pointerType, "mouse", "pointermove.pointerType");
    assert_true(move.isPrimary, "pointermove.isPrimary");
    assert_approx_equals(move.clientX, 70, 0.01, "pointermove.clientX");
    assert_approx_equals(move.clientY, 40, 0.01, "pointermove.clientY");
    assert_approx_equals(move.offsetX, 50, 0.01, "pointermove.offsetX");
    assert_approx_equals(move.offsetY, 20, 0.01, "pointermove.offsetY");
  }, "native pointer movement publishes browser entry order and coordinate fields");

  promise_test(async () => {
    await test_driver.click(target);

    for (const type of ["pointerdown", "mousedown", "pointerup", "mouseup", "click"]) {
      assert_equals(count(type), 1, `${type} must be dispatched exactly once on activation`);
      assert_equals(events(type)[0].target, target, `${type}.target`);
    }
    assert_less_than(observed.findIndex(event => event.type === "pointerdown"),
      observed.findIndex(event => event.type === "mousedown"),
      "pointerdown precedes mousedown");
    assert_less_than(observed.findIndex(event => event.type === "mousedown"),
      observed.findIndex(event => event.type === "pointerup"),
      "mousedown precedes pointerup");
    assert_less_than(observed.findIndex(event => event.type === "pointerup"),
      observed.findIndex(event => event.type === "mouseup"),
      "pointerup precedes mouseup");
    assert_less_than(observed.findIndex(event => event.type === "mouseup"),
      observed.findIndex(event => event.type === "click"),
      "mouseup precedes click");

    assert_equals(events("pointerdown")[0].button, 0, "primary pointerdown.button");
    assert_equals(events("pointerdown")[0].buttons, 1, "primary pointerdown.buttons");
    assert_equals(events("pointerup")[0].button, 0, "primary pointerup.button");
    assert_equals(events("pointerup")[0].buttons, 0, "primary pointerup.buttons");
    assert_equals(events("click")[0].which, 1, "primary click.which");
    assert_equals(events("click")[0].detail, 1, "primary click.detail");
  }, "native primary activation publishes pointer, mouse, and click semantics in order");

  promise_test(async () => {
    await new test_driver.Actions().pointerMove(0, 0, { origin: outside }).send();

    for (const type of ["pointerout", "pointerleave", "mouseout", "mouseleave"]) {
      assert_equals(count(type), 1, `${type} must be dispatched exactly once on exit`);
      assert_equals(events(type)[0].target, target, `${type}.target`);
      assert_equals(events(type)[0].relatedTarget, outside, `${type}.relatedTarget`);
    }
    assert_less_than(observed.findIndex(event => event.type === "pointerout"),
      observed.findIndex(event => event.type === "pointerleave"),
      "pointerout precedes pointerleave");
    assert_less_than(observed.findIndex(event => event.type === "mouseout"),
      observed.findIndex(event => event.type === "mouseleave"),
      "mouseout precedes mouseleave");
  }, "native pointer exit publishes relatedTarget and boundary order");

  promise_test(async () => {
    const parent = document.getElementById("propagation-parent");
    const propagationTarget = document.getElementById("propagation-target");
    const trace = [];
    propagationTarget.addEventListener("mouseup", event => {
      trace.push("first-target");
      event.stopImmediatePropagation();
    });
    propagationTarget.addEventListener("mouseup", () => trace.push("second-target"));
    parent.addEventListener("mouseup", () => trace.push("parent"));
    document.addEventListener("mouseup", () => trace.push("document"));

    await test_driver.click(propagationTarget);

    assert_array_equals(
      trace,
      ["first-target"],
      "stopImmediatePropagation must stop later target and ancestor listeners");
  }, "native mouseup honors stopImmediatePropagation across overlapping recognizers");
</script>
